bmc: Predict halving based on block timestamp #BDK-301
We need to take into account the timestamp of the block instead of
assuming the block has just been mined.

- Fetch both height and timestamp
- Use a channel to send instead of reading an arc
- Predict the halving based on the block timestamp
- Calculate the number of seconds instead of counting down

+++ bmc-display/src/halving_data.rs
@@ -4,7 +4,7 @@
 pub const HALVING_INTERVAL: u32 = 210_000;
 
 /// Average block time in seconds (Bitcoin targets ~10 minutes)
-pub const AVG_BLOCK_TIME_SECS: u64 = 600;
+pub const AVG_BLOCK_TIME_SECS: u32 = 600;
 
 /// Calculate the next halving block height from current height
 #[must_use]
